Changed loop variables for iterating through vectors to be the appropriate type (size_t) to avoid annoying warnings.

+++ SymbolicSummation/AbstractHelpers.hpp
@@ -49,7 +49,7 @@ Expression declareFunction(String name);
 Expression declareFunction(String name, ExprAction action);
 Expression declareSymbol(String name);
 Expression declareSymbol(String name, Expression value);
-Expression declareReal(float value);
+Expression declareReal(double value);
 Expression declareEuclidVector(std::string name, ExprVector components);
 Expression declareEuclidVector(std::string name, std::initializer_list<Expression> components);
 Expression declareMatrix(String matName, ExprMatrix elements);
