+++ SymbolicSummation/AlgebraicHelpers.cpp
@@ -33,7 +33,7 @@ bool isNegative(Expression target) {
         const Add& addObj = dynamic_cast<const Add&>(*target);
         elementsToCheck = addObj.getMembers();
         bool totallyNegative = true;
-        for(int i = 0; i<elementsToCheck.size(); i++) {
+        for(size_t i = 0; i<elementsToCheck.size(); i++) {
             totallyNegative &= isNegative(elementsToCheck[i]);
         }
         return negative != totallyNegative;
@@ -41,7 +41,7 @@ bool isNegative(Expression target) {
     if(targetType == MULTYPE) {
         const Mul& mulObj = dynamic_cast<const Mul&>(*target);
         elementsToCheck = mulObj.getMembers();
-        for(int i = 0; i<elementsToCheck.size(); i++) {
+        for(size_t i = 0; i<elementsToCheck.size(); i++) {
             negative = negative != isNegative(elementsToCheck[i]);
         }
         return negative;
@@ -74,10 +74,10 @@ bool isNegative(Expression target) {
         for (int x = 0; x < n; x++) {
             ExprMatrix submatrix = *new ExprMatrix();
             int subi = 0;
-            for (int i = 1; i < n; i++) {
+            for (size_t i = 1; i < n; i++) {
                 int subj = 0;
                 ExprVector currentRow = *new ExprVector();
-                for (int j = 0; j < n; j++) {
+                for (size_t j =0; j < n; j++) {
                     if (j == x)
                         continue;
                     currentRow.push_back(currMatrix[i][j]);
@@ -106,12 +106,12 @@ Expression matMul(Expression left, Expression right) {
         
         ExprMatrix newElements = *new ExprMatrix();
         String newName = "{";
-        for(int i = 0; i<newDimensions.first; i++) {
+        for(size_t i = 0; i<newDimensions.first; i++) {
             ExprVector currentRow = *new ExprVector();
             newName+="{";
             if(i!=0)
                 newName+=",";
-            for(int j = 0; j<newDimensions.second; j++) {
+            for(size_t j =0; j<newDimensions.second; j++) {
                 Expression result = ZERO;
                 for(int k = 0; k<leftMatrix.dimensions.second; k++) {
                     Expression leftEle = leftMatrix.elements[i][k];
@@ -144,9 +144,9 @@ Expression matMul(Expression left, Expression right) {
     dimensions.first = matTarget.dimensions.second;
     dimensions.second = matTarget.dimensions.first;
     ExprMatrix newElements = *new ExprMatrix();
-    for(int i = 0; i <dimensions.first; i++) {
+    for(size_t i = 0; i <dimensions.first; i++) {
         ExprVector currentRow = *new ExprVector();
-        for(int j = 0; j<dimensions.second; j++) {
+        for(size_t j =0; j<dimensions.second; j++) {
             Expression currElement = matTarget.elements[j][i];
             currentRow.push_back(currElement);
         }
@@ -182,7 +182,7 @@ Expression combineProducts(Expression left, Expression right) {
         if(rightType == MULTYPE) {
             const Mul& rMul = dynamic_cast<const Mul&>(*right);
             ExprVector rightMembers = rMul.getMembers();
-            for(int i = 0; i<rightMembers.size(); i++) {
+            for(size_t i = 0; i<rightMembers.size(); i++) {
                 newMembers.push_back(rightMembers[i]);
             }
         } else {
@@ -197,7 +197,7 @@ Expression combineProducts(Expression left, Expression right) {
         const Mul& rMul = dynamic_cast<const Mul&>(*right);
         ExprVector newMembers = *new ExprVector();
         newMembers.push_back(*new Expression(left));
-        for(int i = 0; i< rMul.getMembers().size(); i++) {
+        for(size_t i = 0; i< rMul.getMembers().size(); i++) {
             newMembers.push_back(rMul.getMembers()[i]);
         }
         Expression result = *new Expression(new Mul(newMembers));
@@ -226,7 +226,7 @@ Expression combineSums(Expression left, Expression right) {
                 Expression tempResult = lAdd.getMembers()[0]+rAdd.getMembers()[0];
                 return tempResult;
             }
-            for(int i = 0; i<rightMembers.size(); i++) {
+            for(size_t i = 0; i<rightMembers.size(); i++) {
                 //Expression nextMember(rightMembers[i]);
                 if(rightMembers[i] == ZERO)
                     continue;
@@ -245,7 +245,7 @@ Expression combineSums(Expression left, Expression right) {
         const Add& rAdd = dynamic_cast<const Add&>(*right);
         ExprVector newMembers = *new ExprVector();
         newMembers.push_back(*new Expression(left));
-        for(int i = 0; i<rAdd.getMembers().size(); i++) {
+        for(size_t i = 0; i<rAdd.getMembers().size(); i++) {
             newMembers.push_back(rAdd.getMembers()[i]);
         }
         Expression result = *new Expression(new Add(newMembers));
@@ -309,12 +309,12 @@ Expression cancelTerms(Expression target) {
     ExprVector newMembers = *new ExprVector();
     std::vector<int> indicesAccountedFor = *new std::vector<int>();
     bool didAnyCancelation = false;
-    for(int i = 0; i<members.size(); i++) {
+    for(size_t i = 0; i<members.size(); i++) {
         if(intVectorContains(indicesAccountedFor, i))
             continue;
         indicesAccountedFor.push_back(i);
         Expression total = members[i];
-        for(int j = i+1; j<members.size(); j++) {
+        for(size_t j =i+1; j<members.size(); j++) {
             if(intVectorContains(indicesAccountedFor, j))
                 continue;
             Expression currMember = members[j];
@@ -478,7 +478,7 @@ Expression simplify(Expression target) {
         if(members.size() == 1)
             return members[0];
         ExprVector newMembers = *new ExprVector();
-        for(int i = 0; i<members.size(); i++) {
+        for(size_t i = 0; i<members.size(); i++) {
             Expression simplified = simplify(members[i]);
             if(simplified == ZERO)
                 return ZERO;
@@ -500,7 +500,7 @@ Expression simplify(Expression target) {
         if(members.size() == 1)
             return *new Expression(target.get());
         ExprVector newMembers = *new ExprVector();
-        for(int i = 0; i<members.size(); i++) {
+        for(size_t i = 0; i<members.size(); i++) {
             Expression simplified = simplify(members[i]);
             if(simplified == ZERO)
                 continue;
@@ -584,8 +584,8 @@ Expression simplify(Expression target) {
 ExprVector cancelCommonFactors(ExprVector targets) {
     ExprVector results = targets;
     ExprVector factors = commonFactors(results);
-    for(int i = 0; i<factors.size(); i++) {
-        for(int j = 0; j<results.size(); j++) {
+    for(size_t i = 0; i<factors.size(); i++) {
+        for(size_t j =0; j<results.size(); j++) {
             results[j] = results[j].factor();
             //bool isContainer = isSubtypeOf(results[j], OPERATORTYPE);
             //if(isContainer)
@@ -740,13 +740,13 @@ ExprVector getFactorsOfInt(Expression factee) {
     const Real& realObj = dynamic_cast<const Real&>(*factee);
     if(!(realObj.value == trunc(realObj.value)))
         return {*new Expression(factee)};
-    int val = realObj.value;
+    int val = (int)realObj.value;
     ExprVector factors = *new ExprVector();
     if(val<0) {
         factors.push_back(MINUSONE);
         val = -val;
     }
-    for(int i = 2; i <= val; i++) {
+    for(size_t i = 2; i <= val; i++) {
         int remain = val%i;
         if(remain == 0) {
             int temp = val;
@@ -766,9 +766,9 @@ ExprVector getFactors(Expression factee) {
     if(type == MULTYPE) {
         const Mul& mulObj = dynamic_cast<const Mul&>(*factee);
         ExprVector factors = *new ExprVector();
-        for(int i = 0; i< mulObj.getMembers().size(); i++) {
+        for(size_t i = 0; i< mulObj.getMembers().size(); i++) {
             ExprVector tempFactors = getFactors(mulObj.getMembers()[i]);
-            for(int j = 0; j< tempFactors.size(); j++)
+            for(size_t j =0; j< tempFactors.size(); j++)
                 //if(!containsElement(factors, tempFactors[j]))
                     factors.push_back(tempFactors[j]);
         }
@@ -801,7 +801,7 @@ ExprVector getFactors(Expression factee) {
         if(exponentType == ADDTYPE) {
             ExprVector factors = *new ExprVector();
             const Add& addObj = dynamic_cast<const Add&>(*exponent);
-            for(int i = 0; i<addObj.getMembers().size(); i++) {
+            for(size_t i = 0; i<addObj.getMembers().size(); i++) {
                 Expression newExpression = *new Expression(new Exp(expObj.getBase(),addObj.getMembers()[i]));
                 factors.push_back(newExpression);
             }
@@ -818,7 +818,7 @@ ExprVector commonFactors(ExprVector terms) {
     ExprVector commonFactors = firstTermFactors;
     if(terms.size() == 1)
         return firstTermFactors;
-    for(int i = 1; i< terms.size(); i++) {
+    for(size_t i = 1; i< terms.size(); i++) {
         ExprVector nextFactors = terms[i].getFactors();
         commonFactors = setIntersect(commonFactors, nextFactors);
     }
@@ -830,13 +830,13 @@ ExprVector commonFactors(ExprVector terms) {
     if(terms.size()<2)
         return firstTermFactors;
     std::vector<std::vector<int>> indicesAlreadyCounted = *new std::vector<std::vector<int>>();
-    for(int i = 1; i<terms.size(); i++) {
+    for(size_t i = 1; i<terms.size(); i++) {
         ExprVector currTermFactors = getFactors(terms[i]);
         ithTermFactors.push_back(currTermFactors);
         indicesAlreadyCounted.push_back(*new std::vector<int>());
     }
     
-    for(int i = 0; i<firstTermFactors.size(); i++) {
+    for(size_t i = 0; i<firstTermFactors.size(); i++) {
         bool isCommonFactor = true;
         for(int j=0; j<(ithTermFactors).size(); j++) {
             bool containsFactor = false;
@@ -893,7 +893,7 @@ Expression combineTermsDifferingByCoefficientsAdditively(Expression left, Expres
     targets = generateExprVector({*new Expression(left),*new Expression(right)});
     ExprVector results = cancelCommonFactors(targets);//cancelCommonFactors(targets);
     Expression inCommon = ONE;
-    for(int i = 0; i< factors.size(); i++) {
+    for(size_t i = 0; i< factors.size(); i++) {
         inCommon = inCommon * factors[i];
     }
     Expression combinedCoeffs = combineTermsDifferingByCoefficientsAdditively(results[0], results[1]);
@@ -908,8 +908,8 @@ Expression performActions(Expression target) {
     if(isTypeSimilarTo(target, MATRIXTYPE)) {
         const Matrix& matObj = dynamic_cast<const Matrix&>(*target);
         ExprMatrix elements = *new ExprMatrix(matObj.elements);
-        for(int i = 0; i<matObj.dimensions.first; i++) {
-            for (int j = 0; j<matObj.dimensions.second; j++) {
+        for(size_t i = 0; i<matObj.dimensions.first; i++) {
+            for (size_t j =0; j<matObj.dimensions.second; j++) {
                 if(elements[i][j] == target) {
                     elements[i][j] = performActions(elements[i][j]);
                 }
@@ -953,7 +953,7 @@ Expression performActions(Expression target) {
         const Mul& mulObj = dynamic_cast<const Mul&>(*target);
         elementsToModify = mulObj.getMembers();
     }
-    for(int i = 0; i< elementsToModify.size(); i++) {
+    for(size_t i = 0; i< elementsToModify.size(); i++) {
         Expression moddedEle = performActions(elementsToModify[i]);
         newElements.push_back(moddedEle);
     }
@@ -971,8 +971,8 @@ Expression performActionsOn(Expression target, Expression var) {
     if(isTypeSimilarTo(target, MATRIXTYPE)) {
         const Matrix& matObj = dynamic_cast<const Matrix&>(*target);
         ExprMatrix elements = *new ExprMatrix(matObj.elements);
-        for(int i = 0; i<matObj.dimensions.first; i++) {
-            for (int j = 0; j<matObj.dimensions.second; j++) {
+        for(size_t i = 0; i<matObj.dimensions.first; i++) {
+            for (size_t j =0; j<matObj.dimensions.second; j++) {
                 if(elements[i][j] == target) {
                     elements[i][j] = performActionsOn(elements[i][j],var);
                 }
@@ -1016,7 +1016,7 @@ Expression performActionsOn(Expression target, Expression var) {
         const Mul& mulObj = dynamic_cast<const Mul&>(*target);
         elementsToModify = mulObj.getMembers();
     }
-    for(int i = 0; i< elementsToModify.size(); i++) {
+    for(size_t i = 0; i< elementsToModify.size(); i++) {
         Expression moddedEle = performActionsOn(elementsToModify[i], var);
         newElements.push_back(moddedEle);
     }
@@ -1034,8 +1034,8 @@ Expression insertAsVariable(Expression target, Expression var) {
     if(isTypeSimilarTo(target, MATRIXTYPE)) {
         const Matrix& matObj = dynamic_cast<const Matrix&>(*target);
         ExprMatrix elements = *new ExprMatrix(matObj.elements);
-        for(int i = 0; i<matObj.dimensions.first; i++) {
-            for (int j = 0; j<matObj.dimensions.second; j++) {
+        for(size_t i = 0; i<matObj.dimensions.first; i++) {
+            for (size_t j =0; j<matObj.dimensions.second; j++) {
                 if(elements[i][j] == target) {
                     elements[i][j] = insertAsVariable(elements[i][j],var);
                 }
@@ -1079,7 +1079,7 @@ Expression insertAsVariable(Expression target, Expression var) {
         const Mul& mulObj = dynamic_cast<const Mul&>(*target);
         elementsToModify = mulObj.getMembers();
     }
-    for(int i = 0; i< elementsToModify.size(); i++) {
+    for(size_t i = 0; i< elementsToModify.size(); i++) {
         Expression moddedEle = insertAsVariable(elementsToModify[i], var);
         newElements.push_back(moddedEle);
     }
@@ -1101,8 +1101,8 @@ Expression substitute(Expression source, Expression target, Expression value) {
         const Matrix& matObj = dynamic_cast<const Matrix&>(*source);
         ExprMatrix elements = *new ExprMatrix(matObj.elements);
         bool foundTarget = false;
-        for(int i = 0; i<matObj.dimensions.first; i++) {
-            for (int j = 0; j<matObj.dimensions.second; j++) {
+        for(size_t i = 0; i<matObj.dimensions.first; i++) {
+            for (size_t j =0; j<matObj.dimensions.second; j++) {
                 if(elements[i][j] == target) {
                     elements[i][j] = valueExpr;
                     foundTarget = true;
@@ -1194,7 +1194,7 @@ Expression substitute(Expression source, Expression target, Expression value) {
     }
     
     bool moddedAnyElements = false;
-    for(int i = 0; i< elementsToModify.size(); i++) {
+    for(size_t i = 0; i< elementsToModify.size(); i++) {
         Expression moddedEle = substitute(elementsToModify[i], target, value);
         if(moddedEle.getTypeHash() == NULLTYPE) {
             newElements.push_back(elementsToModify[i]);
@@ -1216,7 +1216,7 @@ Expression substitute(Expression source, Expression target, Expression value) {
 
 Expression substitute(Expression source, std::vector<std::pair<Expression,Expression>> substitutions) {
     Expression result = *new Expression(source);
-    for(int i = 0; substitutions.size(); i++) {
+    for(size_t i = 0; substitutions.size(); i++) {
         Expression temp = substitute(result, substitutions[i].first, substitutions[i].second);
         if(temp.getTypeHash() != NULLTYPE)
             result = temp;
